Correções em algumas configs

+++ api/src/main/kotlin/com/api/application/task/exceptions/TaskException.kt
@@ -1,13 +1,13 @@
 package com.api.application.task.exceptions
 
 sealed class TaskExceptions(message: String): Exception(message) {
-    abstract val taskID: Int?
+    abstract val taskID: Long?
 }
 
 data class TaskNotFoundException (
-    override val taskID: Int?
+    override val taskID: Long?
 ): TaskExceptions("Projeto $taskID não encontrado!")
 
 data class TaskNotInsertException (
-    override val taskID: Int?
+    override val taskID: Long?
 ): TaskExceptions("Não foi possivel inserir o Projeto de ID: $taskID!")
